Add endpoint for the Nextstrain CLI standalone archive installers
Currently these resolve through redirects to GitHub where the installers
are served directly out of the CLI's repo.¹  This will make maintenance
and updates easier, and the indirection of the nextstrain.org endpoints
allows us to swap things around behind the scenes as necessary later.

Planned to be used by forthcoming updates to a) our installation docs²
and b) the self-upgrade instructions of the CLI itself (when already
running a standalone install).

¹ https://github.com/nextstrain/cli/pull/217
² https://github.com/nextstrain/docs.nextstrain.org/pull/123