Upgrade @aws-sdk/client-cognito-identity-provider so we can import exception classes
Otherwise we have to do string comparison on the exception names instead
of cleaner instanceof checks.

I was confused for a bit because the documentation says you can import
these classes but that wasn't working in practice when I first tried it.
Turns out we require 3.52.0 and the classes weren't exported until
3.53.0, just one release later!