Use Group -> Source over GroupSource -> Group
Previously, only the Group was a property of GroupSource. Having the
GroupSource as a property of Group allows for more intuitive access.

However, maintaining both of these properties results in a cyclic
dependency when require statements are used at the top of each file.
Choosing to keep the require at the top of Group rather than GroupSource
because Group -> source makes more sense in the larger hierarchical
model.

Also, set req.context.group on /groups/:groupName.