Remove authz assertion when setting the group
In the previous commit, the group is set for all requests that contain
the group's URL namespace (/groups/:groupName). This includes datasets
and narratives where the first level of authorization was the group
source authorization in setSource, which includes the Visibility.Public
tag to allow public Read access where appropriate.

It does not make sense to set Visibility.Public on the Group itself,
since things like reading members lists on public groups should still be
gated by authorization.